requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
Appium-Python-Client==1.0.1
attrs==19.3.0
helpers==0.2.0
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
selenium==3.141.0
six==1.15.0
softest==1.2.0.0
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
